Fix Python 3.10 compat and pre-commit issues
asyncio.Barrier is 3.11+, so use an Event + counter instead. Split the
concurrency tests into their own file to stay under the 500-line loq
limit, and fix pyright suppression comments.
